% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\docType{data}
\name{dataset_noaa_raw}
\alias{dataset_noaa_raw}
\title{dataset obtained from the U.S. National Oceanographic and Atmospheric Administration (NOAA)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5958 rows and 47 columns.}
\usage{
dataset_noaa_raw
}
\description{
dataset obtained from the U.S. National Oceanographic and Atmospheric Administration (NOAA)
}
\keyword{datasets}
