% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean dataset obtained from the U.S. National Oceanographic and Atmospheric Administration (NOAA)}
\usage{
eq_clean_data(raw)
}
\arguments{
\item{raw}{dataset fo earthquakes at data.frame format}
}
\value{
Returns cleaned and formatted data.frame
}
\description{
Function takes raw dataset as data.frame which contains DATE, YEAR, MONTH, DAY, LATITUDE, LONGITUDE,
 TOTAL_DEATHS,EQ_PRIMARY,LOCATION_NAME varaibles.
 Dataset can be obtained from NOAA's
\href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{Significant
EarthQuake Database}
}
\examples{
eq_clean_data(d)

}
