% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Geom Timeline Label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\value{
ggplot layer with labels
}
\description{
Function plots Labels for earthquakes.
}
\examples{
# d\%>\%subset(COUNTRY=="USA" | COUNTRY=="CHINA")\%>\%subset(DATE>=as.Date("2000-01-01"))\%>\%
# ggplot(aes(DATE, COUNTRY, colour = DEATHS, size = EQ_PRIMARY, label = LOCATION_NAME)) +
#  geom_timeline()+ geom_timeline_label(n_max=5)+
#   guides(size = guide_legend("Richter scale value")) +
#   guides(colour = guide_colorbar("# Deaths", label.theme = element_text(angle = 45, size = 8 ))) +
#   ylab("") +
#   theme_classic() +
#   theme(legend.position="bottom", axis.line.y = element_blank())
}
