% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Location clean for dataset obtained from the U.S. National Oceanographic and Atmospheric Administration (NOAA)}
\usage{
eq_location_clean(raw)
}
\arguments{
\item{raw}{dataset fo earthquakes at data.frame format}
}
\value{
dataframe with trimmed LOCATION_NAME field
}
\description{
Dataset contains varaible LOCATION_NAME that should be trimmed. This function deletes names of countries and regions which
in LOCATION_NAME placed before ":"
}
\examples{
\dontrun{
eq_location_clean(dataset_noaa_raw)
}
}
