% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Map of Significant Earthquakes}
\usage{
eq_map(data, annot_col = "popup text")
}
\arguments{
\item{data}{Dataset of earthquakes. For cleaning use \code{\link{eq_clean_data}}}

\item{annot_col}{Use this parameter for popup annotation with specific varaible}
}
\value{
Leaflet map
}
\description{
Creates map with circle points of earthqakes with popup annotation
}
\examples{
d<-dataset_noaa_raw
d<-dplyr::filter(eq_clean_data(d), COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000)
eq_map(d)
# For Date annotation popup
eq_map(d, annot_col="DATE")

}
